+++ src/test/java/io/github/annusshka/MatrixTest.java
@@ -1,8 +1,8 @@
 package io.github.annusshka;
 
 import io.github.annusshka.Math.Matrix.Matrix;
-import io.github.annusshka.Math.Matrix.Matrix3d;
-import io.github.annusshka.Math.Matrix.Matrix4d;
+import io.github.annusshka.Math.Matrix.Matrix3f;
+import io.github.annusshka.Math.Matrix.Matrix4f;
 import io.github.annusshka.Math.Vector.*;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -13,104 +13,104 @@ public class MatrixTest {
     static final float EPS = 1e-7f;
     @Test
     public void createIdentityMatrix() {
-        Matrix3d matrix3x = new Matrix3d(new float[]{5.4f, 0, 0, 0, 5.4f, 0, 0, 0, 5.4f});
+        Matrix3f matrix3x = new Matrix3f(new float[]{5.4f, 0, 0, 0, 5.4f, 0, 0, 0, 5.4f});
         Assertions.assertThat(matrix3x.createIdentityMatrix(5.4f).getVector()).
                 isEqualTo(matrix3x.getVector());
 
-        Matrix4d matrix4x = new Matrix4d(new float[]{3, 0, 0, 0, 0, 3, 0, 0, 0, 0, 3, 0, 0, 0, 0, 3});
+        Matrix4f matrix4x = new Matrix4f(new float[]{3, 0, 0, 0, 0, 3, 0, 0, 0, 0, 3, 0, 0, 0, 0, 3});
         Assertions.assertThat(matrix4x.createIdentityMatrix(3).getVector()).
                 isEqualTo(matrix4x.getVector());
     }
 
     @Test
     public void isIdentityMatrix() {
-        Matrix3d matrix3x1 = new Matrix3d(new float[]{-0.5f, 0, 0, 0, -0.5f, 0, 0, 0, -0.5f});
+        Matrix3f matrix3x1 = new Matrix3f(new float[]{-0.5f, 0, 0, 0, -0.5f, 0, 0, 0, -0.5f});
         Assertions.assertThat(Matrix.isIdentityMatrix(matrix3x1, EPS)).isEqualTo(true);
 
-        Matrix3d matrix3x2 = new Matrix3d(new float[]{-0.5f, 6, 0, 0, -0.5f, -0.5f, 0, 0, -0.5f});
+        Matrix3f matrix3x2 = new Matrix3f(new float[]{-0.5f, 6, 0, 0, -0.5f, -0.5f, 0, 0, -0.5f});
         Assertions.assertThat(Matrix.isIdentityMatrix(matrix3x2, EPS)).isEqualTo(false);
 
-        Matrix3d matrix3x3 = new Matrix3d(new float[]{0, 0, 0, 0, 0, 0, 0, 0, 0});
+        Matrix3f matrix3x3 = new Matrix3f(new float[]{0, 0, 0, 0, 0, 0, 0, 0, 0});
         Assertions.assertThat(Matrix.isIdentityMatrix(matrix3x3, EPS)).isEqualTo(false);
 
-        Matrix4d matrix4x1 = new Matrix4d(new float[]{3, 0, 0, 0, 0, 3, 0, 0, 0, 0, 3, 0, 0, 0, 0, 3});
+        Matrix4f matrix4x1 = new Matrix4f(new float[]{3, 0, 0, 0, 0, 3, 0, 0, 0, 0, 3, 0, 0, 0, 0, 3});
         Assertions.assertThat(Matrix.isIdentityMatrix(matrix4x1, EPS)).isEqualTo(true);
 
-        Matrix4d matrix4x2 = new Matrix4d(new float[]{3, 0, 0.4f, 0, 0, 3, 0, 0, 6.5f, 0, 3, 0, 0, 0, 0, 3});
+        Matrix4f matrix4x2 = new Matrix4f(new float[]{3, 0, 0.4f, 0, 0, 3, 0, 0, 6.5f, 0, 3, 0, 0, 0, 0, 3});
         Assertions.assertThat(Matrix.isIdentityMatrix(matrix4x2, EPS)).isEqualTo(false);
     }
 
     @Test
     public void getZeroMatrix() {
-        Assertions.assertThat(Objects.requireNonNull(Matrix3d.getZeroMatrix()).getVector()).
+        Assertions.assertThat(Objects.requireNonNull(Matrix3f.getZeroMatrix()).getVector()).
                 isEqualTo(new float[]{0, 0, 0, 0, 0, 0, 0, 0, 0});
     }
 
     @Test
     public void transposeMatrix() {
-        Matrix3d matrix3x = new Matrix3d(new float[]{0, 1.1f, -3, 0, -4.5f, 7.3f, 6, 0.78f, 1});
+        Matrix3f matrix3x = new Matrix3f(new float[]{0, 1.1f, -3, 0, -4.5f, 7.3f, 6, 0.78f, 1});
         Assertions.assertThat(Matrix.transposeMatrix(matrix3x).getVector()).
                 isEqualTo(new float[]{0, 0, 6, 1.1f, -4.5f, 0.78f, -3, 7.3f, 1});
 
-        Matrix4d matrix4x = new Matrix4d(new float[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15});
+        Matrix4f matrix4x = new Matrix4f(new float[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15});
         Assertions.assertThat(Matrix.transposeMatrix(matrix4x).getVector()).
                 isEqualTo(new float[]{0, 4, 8, 12, 1, 5, 9, 13, 2, 6, 10, 14, 3, 7, 11, 15});
     }
 
     @Test
     public void sumMatrix() throws Matrix.MatrixException {
-        Matrix4d matrix4x1 = new Matrix4d(new float[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15});
-        Matrix4d matrix4x2 = new Matrix4d(
+        Matrix4f matrix4x1 = new Matrix4f(new float[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15});
+        Matrix4f matrix4x2 = new Matrix4f(
                 new float[]{0, -1, -2, -3, -4, -5, -6, -7, -8, -9, -10, -11, -12, -13, -14, -15});
 
-        Assertions.assertThat(Matrix.sumMatrix(matrix4x1, Matrix4d.getZeroMatrix()).getVector()).
+        Assertions.assertThat(Matrix.sumMatrix(matrix4x1, Matrix4f.getZeroMatrix()).getVector()).
                 isEqualTo(new float[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15});
         Assertions.assertThat(Matrix.sumMatrix(matrix4x1, matrix4x1).getVector()).
                 isEqualTo(new float[]{0, 2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30});
         Assertions.assertThat(Matrix.sumMatrix(matrix4x1, matrix4x2).getVector()).
-                isEqualTo(Matrix4d.getZeroMatrix().getVector());
+                isEqualTo(Matrix4f.getZeroMatrix().getVector());
     }
 
     @Test
     public void minusMatrix() throws Matrix.MatrixException {
-        Matrix4d matrix4x1 = new Matrix4d(new float[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15});
-        Matrix4d matrix4x3 = new Matrix4d(
+        Matrix4f matrix4x1 = new Matrix4f(new float[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15});
+        Matrix4f matrix4x3 = new Matrix4f(
                 new float[]{0, -1, -2, -3, -4, -5, -6, -7, -8, -9, -10, -11, -12, -13, -14, -15});
 
-        Assertions.assertThat(Matrix.minusMatrix(matrix4x1, Matrix4d.getZeroMatrix()).getVector()).
+        Assertions.assertThat(Matrix.minusMatrix(matrix4x1, Matrix4f.getZeroMatrix()).getVector()).
                 isEqualTo(new float[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15});
         Assertions.assertThat(Matrix.minusMatrix(matrix4x1, matrix4x1).getVector()).
-                isEqualTo(Matrix4d.getZeroMatrix().getVector());
+                isEqualTo(Matrix4f.getZeroMatrix().getVector());
         Assertions.assertThat(Matrix.minusMatrix(matrix4x1, matrix4x3).getVector()).
                 isEqualTo(new float[]{0, 2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30});
     }
 
     @Test
     public void multiplicateOnValue() {
-        Matrix3d matrix3x1 = new Matrix3d(new float[]{1, 2, 3, 4, 3, 2, 1, 0, 0});
+        Matrix3f matrix3x1 = new Matrix3f(new float[]{1, 2, 3, 4, 3, 2, 1, 0, 0});
         matrix3x1.multiplicateOnValue(3);
         Assertions.assertThat(matrix3x1.getVector()).isEqualTo(new float[]{3, 6, 9, 12, 9, 6, 3, 0, 0});
 
-        Matrix3d matrix3x2 = new Matrix3d(new float[]{1, 2, 3, 4, 3, 2, 1, 0, 0});
+        Matrix3f matrix3x2 = new Matrix3f(new float[]{1, 2, 3, 4, 3, 2, 1, 0, 0});
         matrix3x2.multiplicateOnValue(0);
         Assertions.assertThat(matrix3x2.getVector()).isEqualTo(new float[matrix3x2.getLength()]);
 
-        Matrix3d matrix3x3 = new Matrix3d(new float[]{1, 2, 3, 4, 3, -2, -1, 0, 0});
-        Matrix3d result3x3 = new Matrix3d(new float[]{-0.73f, -1.46f, -2.19f, -2.92f, -2.19f, 1.46f, 0.73f, 0, 0});
+        Matrix3f matrix3x3 = new Matrix3f(new float[]{1, 2, 3, 4, 3, -2, -1, 0, 0});
+        Matrix3f result3x3 = new Matrix3f(new float[]{-0.73f, -1.46f, -2.19f, -2.92f, -2.19f, 1.46f, 0.73f, 0, 0});
         matrix3x3.multiplicateOnValue(-0.73f);
         Assertions.assertThat(matrix3x3.isEqualMatrix(result3x3)).isEqualTo(true);
 
-        Matrix4d matrix4x1 = new Matrix4d(new float[]{1, 2, 3, 4, 3, 2, 1, 0, 0, 4, 3, 2, 1, 8, 4, 3});
+        Matrix4f matrix4x1 = new Matrix4f(new float[]{1, 2, 3, 4, 3, 2, 1, 0, 0, 4, 3, 2, 1, 8, 4, 3});
         matrix4x1.multiplicateOnValue(3);
         Assertions.assertThat(matrix4x1.getVector()).
                 isEqualTo(new float[]{3, 6, 9, 12, 9, 6, 3, 0, 0, 12, 9, 6, 3, 24, 12, 9});
 
-        Matrix4d matrix4x2 = new Matrix4d(new float[]{1, 2, 3, 4, 3, 2, 1, 0, 0, 4, 3, 2, 1, 8, 4, 3});
+        Matrix4f matrix4x2 = new Matrix4f(new float[]{1, 2, 3, 4, 3, 2, 1, 0, 0, 4, 3, 2, 1, 8, 4, 3});
         matrix4x2.multiplicateOnValue(0);
         Assertions.assertThat(matrix4x2.getVector()).isEqualTo(new float[matrix4x2.getLength()]);
 
-        Matrix4d matrix4x3 = new Matrix4d(new float[]{1, 2, 3, 4, 3, -2, -1, 0, 0, 4, 3, 2, 1, 0, 4, 3});
-        Matrix4d result4x3 = new Matrix4d(new float[]
+        Matrix4f matrix4x3 = new Matrix4f(new float[]{1, 2, 3, 4, 3, -2, -1, 0, 0, 4, 3, 2, 1, 0, 4, 3});
+        Matrix4f result4x3 = new Matrix4f(new float[]
                 {-0.73f, -1.46f, -2.19f, -2.92f, -2.19f, 1.46f, 0.73f, 0, 0, -2.92f, -2.19f, -1.46f, -0.73f,
                         0, -2.92f, -2.19f});
         matrix4x3.multiplicateOnValue(-0.73f);
@@ -119,12 +119,12 @@ public class MatrixTest {
 
     @Test
     public void divideOnValue() throws Matrix.MatrixException {
-        Matrix3d matrix3x1 = new Matrix3d(new float[]{3, 6, 9, 12, 9, 6, 3, 0, 0});
+        Matrix3f matrix3x1 = new Matrix3f(new float[]{3, 6, 9, 12, 9, 6, 3, 0, 0});
         matrix3x1.divideOnValue(3);
         Assertions.assertThat(matrix3x1.getVector()).
                 isEqualTo(new float[]{1, 2, 3, 4, 3, 2, 1, 0, 0});
 
-        Matrix3d matrix3x2 = new Matrix3d(new float[]{3, 6, 9, 12, 9, 6, 3, 0, 0});
+        Matrix3f matrix3x2 = new Matrix3f(new float[]{3, 6, 9, 12, 9, 6, 3, 0, 0});
         Throwable thrown2 = Assertions.catchThrowable(() -> {
             matrix3x2.divideOnValue(0);
         });
@@ -132,8 +132,8 @@ public class MatrixTest {
         Assertions.assertThat(thrown2.getMessage()).isNotBlank();
         Assertions.assertThat(thrown2.getMessage()).isEqualTo("Division by zero");
 
-        Matrix4d matrix4x1 = new Matrix4d(new float[]{3, 6, 9, 12, 9, 6, -3, 0, 0, 2, 3, -4, 5, 1, 7, 3});
-        Matrix4d result4x1 = new Matrix4d(new float[]
+        Matrix4f matrix4x1 = new Matrix4f(new float[]{3, 6, 9, 12, 9, 6, -3, 0, 0, 2, 3, -4, 5, 1, 7, 3});
+        Matrix4f result4x1 = new Matrix4f(new float[]
                 {-15, -30, -45, -60, -45, -30, 15, 0, 0, -10, -15, 20, -25, -5, -35, -15});
         matrix4x1.divideOnValue(-0.2f);
         Assertions.assertThat(matrix4x1.isEqualMatrix(result4x1)).isEqualTo(true);
@@ -141,13 +141,13 @@ public class MatrixTest {
 
     @Test
     public void multiplicateOnVector() throws Matrix.MatrixException {
-        Matrix3d matrix3x = new Matrix3d(new float[]{1, 2, 3, 4, 5, 6, 7, 8, 9});
-        Vector3d vector3f = new Vector3d(new float[]{0.0f, -0.5f, 1.7f});
-        Vector3d result3f = new Vector3d(new float[]{4.1f, 7.7f, 11.3f});
+        Matrix3f matrix3x = new Matrix3f(new float[]{1, 2, 3, 4, 5, 6, 7, 8, 9});
+        Vector3f vector3f = new Vector3f(new float[]{0.0f, -0.5f, 1.7f});
+        Vector3f result3f = new Vector3f(new float[]{4.1f, 7.7f, 11.3f});
         Assertions.assertThat(matrix3x.multiplicateOnVector(vector3f).isEqual(result3f)).isEqualTo(true);
 
-        Matrix4d matrix4x = new Matrix4d(new float[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15});
-        Vector4d vector4f = new Vector4d(new float[]{0, 1, 2, 3});
+        Matrix4f matrix4x = new Matrix4f(new float[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15});
+        Vector4f vector4f = new Vector4f(new float[]{0, 1, 2, 3});
         Assertions.assertThat(matrix4x.multiplicateOnVector(vector4f).getVector()).
                 isEqualTo(new float[]{14, 38, 62, 86});
         Assertions.assertThat(matrix4x.multiplicateOnVector(matrix4x.getZeroVector(matrix4x.getSize())).getVector()).
@@ -163,13 +163,13 @@ public class MatrixTest {
 
     @Test
     public void multiplicateMatrices() throws Matrix.MatrixException {
-        Matrix3d matrix3x1 = new Matrix3d(new float[]{0, 1, 2, 3, 4, 5, 6, 7, 8});
-        Matrix3d matrix3x2 = new Matrix3d(new float[]{9, 1, 2, 3, 4, 5, 6, 7, 8});
+        Matrix3f matrix3x1 = new Matrix3f(new float[]{0, 1, 2, 3, 4, 5, 6, 7, 8});
+        Matrix3f matrix3x2 = new Matrix3f(new float[]{9, 1, 2, 3, 4, 5, 6, 7, 8});
 
         Assertions.assertThat(Matrix.multiplicateMatrices(matrix3x1, matrix3x2).getVector()).
                 isEqualTo(new float[]{15, 18, 21, 69, 54, 66, 123, 90, 111});
 
-        Matrix4d matrix4x1 = new Matrix4d(new float[]{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16});
+        Matrix4f matrix4x1 = new Matrix4f(new float[]{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16});
         Throwable thrown2 = Assertions.catchThrowable(() -> {
             Matrix.multiplicateMatrices(matrix3x1, matrix4x1);
         });
@@ -177,9 +177,9 @@ public class MatrixTest {
         Assertions.assertThat(thrown2.getMessage()).isNotBlank();
         Assertions.assertThat(thrown2.getMessage()).isEqualTo("Different sizes can't be multiplicated");
 
-        Matrix4d matrix4x2 = new Matrix4d(
+        Matrix4f matrix4x2 = new Matrix4f(
                 new float[]{0, 1.5f, 0, -3.8f, 4, 0, 6, -1.7f, 8, 9, 1, 11, -12, 0, 4, 0.5f});
-        Matrix4d result = new Matrix4d(
+        Matrix4f result = new Matrix4f(
                 new float[]{-16, 28.5f, 31, 27.8f, -16, 70.5f, 75, 51.8f, -16, 112.5f, 119, 75.8f,
                         -16, 154.5f, 163, 99.8f});
         Assertions.assertThat(Matrix.multiplicateMatrices(matrix4x1, matrix4x2).isEqualMatrix(result)).
@@ -188,20 +188,20 @@ public class MatrixTest {
 
     @Test
     public void getMatrixDeterminant() {
-        Matrix3d matrix3x1 = new Matrix3d(new float[]{1, -2, 3, 4, 0, 6, -7, 8, 9});
-        Assertions.assertThat(Matrix3d.getMatrixDeterminant(matrix3x1)).isEqualTo(204);
+        Matrix3f matrix3x1 = new Matrix3f(new float[]{1, -2, 3, 4, 0, 6, -7, 8, 9});
+        Assertions.assertThat(Matrix3f.getMatrixDeterminant(matrix3x1)).isEqualTo(204);
 
-        Matrix3d matrix3x2 = new Matrix3d(new float[]{1, 0, 3, 4, 0, 6, -7, 0, 9});
-        Assertions.assertThat(Matrix3d.getMatrixDeterminant(matrix3x2)).isEqualTo(0);
+        Matrix3f matrix3x2 = new Matrix3f(new float[]{1, 0, 3, 4, 0, 6, -7, 0, 9});
+        Assertions.assertThat(Matrix3f.getMatrixDeterminant(matrix3x2)).isEqualTo(0);
 
-        Matrix4d matrix4x = new Matrix4d(new float[]{10, 0, 0, 0, 0, 4, 5, 2, 6, 2, 3, 3, 4, 1, 2, 1});
-        Assertions.assertThat(Matrix4d.getMatrixDeterminant(matrix4x)).isEqualTo(-50);
+        Matrix4f matrix4x = new Matrix4f(new float[]{10, 0, 0, 0, 0, 4, 5, 2, 6, 2, 3, 3, 4, 1, 2, 1});
+        Assertions.assertThat(Matrix4f.getMatrixDeterminant(matrix4x)).isEqualTo(-50);
     }
 
     @Test
     public void getDeterminant() {
-        Matrix3d matrix3x1 = new Matrix3d(new float[]{2, 5, 7, 6, 3, 4, 5, -2, -3});
-        Matrix3d matrix3x2 = new Matrix3d(new float[]{2, 0, 7, 6, 0, 4, 5, 0, -3});
+        Matrix3f matrix3x1 = new Matrix3f(new float[]{2, 5, 7, 6, 3, 4, 5, -2, -3});
+        Matrix3f matrix3x2 = new Matrix3f(new float[]{2, 0, 7, 6, 0, 4, 5, 0, -3});
 
         Assertions.assertThat(matrix3x1.getDeterminant()).isEqualTo(-1);
         Assertions.assertThat(matrix3x2.getDeterminant()).isEqualTo(0);
@@ -209,14 +209,14 @@ public class MatrixTest {
 
     @Test
     public void getInverseMatrix() throws Exception {
-        Matrix3d matrix3x1 = new Matrix3d(new float[]{2, 5, 7, 6, 3, 4, 5, -2, -3});
-        Matrix3d inverseMatrix3x1 = new Matrix3d(new float[]{1, -1, 1, -38, 41, -34, 27, -29, 24});
+        Matrix3f matrix3x1 = new Matrix3f(new float[]{2, 5, 7, 6, 3, 4, 5, -2, -3});
+        Matrix3f inverseMatrix3x1 = new Matrix3f(new float[]{1, -1, 1, -38, 41, -34, 27, -29, 24});
         Assertions.assertThat(Matrix.getInverseMatrix(matrix3x1).isEqualMatrix(inverseMatrix3x1)).isEqualTo(true);
         Assertions.assertThat(Matrix.multiplicateMatrices(
-                new Matrix3d(new float[]{2, 5, 7, 6, 3, 4, 5, -2, -3}), inverseMatrix3x1).isEqualMatrix(matrix3x1)).
+                new Matrix3f(new float[]{2, 5, 7, 6, 3, 4, 5, -2, -3}), inverseMatrix3x1).isEqualMatrix(matrix3x1)).
                 isEqualTo(true);
 
-        Matrix3d matrix3x2 = new Matrix3d(new float[]{2, 0, 7, 6, 0, 4, 5, 0, -3});
+        Matrix3f matrix3x2 = new Matrix3f(new float[]{2, 0, 7, 6, 0, 4, 5, 0, -3});
         Throwable thrown = Assertions.catchThrowable(() -> {
             Matrix.getInverseMatrix(matrix3x2);
         });
@@ -227,25 +227,25 @@ public class MatrixTest {
 
     @Test
     public void solutionByGaussMethod() throws Matrix.MatrixException {
-        Matrix3d matrix3x1 = new Matrix3d(new float[]{3, 2, -5, 2, -1, 3, 1, 2, -1});
-        Vector3d vector3f1 = new Vector3d(new float[]{-1, 13, 9});
+        Matrix3f matrix3x1 = new Matrix3f(new float[]{3, 2, -5, 2, -1, 3, 1, 2, -1});
+        Vector3f vector3f1 = new Vector3f(new float[]{-1, 13, 9});
         Assertions.assertThat(Matrix.solutionByGaussMethod(matrix3x1, vector3f1).getVector()).
                 isEqualTo(new float[]{3, 5, 4});
 
-        Matrix3d matrix3x = new Matrix3d(new float[]{2, 1, -1, -3, -1, 2, -2, 1, 2});
-        Vector3d vector3f = new Vector3d(new float[]{8, -11, -3});
+        Matrix3f matrix3x = new Matrix3f(new float[]{2, 1, -1, -3, -1, 2, -2, 1, 2});
+        Vector3f vector3f = new Vector3f(new float[]{8, -11, -3});
         Assertions.assertThat(Matrix.solutionByGaussMethod(matrix3x, vector3f).getVector()).
                 isEqualTo(new float[]{2, 3, -1});
 
         //Бесконечно много решений, но найдём частное решение
-        Matrix3d matrix3x2 = new Matrix3d(new float[]{1, 1, -1, 3, 2, -5, 3, 1, -7});
-        Vector3d vector3f2 = new Vector3d(new float[]{4, 7, 2});
+        Matrix3f matrix3x2 = new Matrix3f(new float[]{1, 1, -1, 3, 2, -5, 3, 1, -7});
+        Vector3f vector3f2 = new Vector3f(new float[]{4, 7, 2});
         Assertions.assertThat(Matrix.solutionByGaussMethod(matrix3x2, vector3f2).getVector()).
                 isEqualTo(new float[]{2, 3, 1});
 
         //Нет решений
-        Matrix3d matrix3x3 = new Matrix3d(new float[]{0, 1, 0, 0, 1, 0, 0, 0, 0});
-        Vector3d vector3f3 = new Vector3d(new float[]{8, 6, 3});
+        Matrix3f matrix3x3 = new Matrix3f(new float[]{0, 1, 0, 0, 1, 0, 0, 0, 0});
+        Vector3f vector3f3 = new Vector3f(new float[]{8, 6, 3});
         Throwable thrown = Assertions.catchThrowable(() -> {
             Matrix.solutionByGaussMethod(matrix3x3, vector3f3);
         });
@@ -253,8 +253,8 @@ public class MatrixTest {
         Assertions.assertThat(thrown.getMessage()).isNotBlank();
         Assertions.assertThat(thrown.getMessage()).isEqualTo("There are no solutions");
 
-        Matrix3d matrix3x4 = new Matrix3d(new float[]{1, 0, 0, 0, 0, 0, 2, 0, 0});
-        Vector3d vector3f4 = new Vector3d(new float[]{8, 0, 0});
+        Matrix3f matrix3x4 = new Matrix3f(new float[]{1, 0, 0, 0, 0, 0, 2, 0, 0});
+        Vector3f vector3f4 = new Vector3f(new float[]{8, 0, 0});
         Throwable thrown2 = Assertions.catchThrowable(() -> {
             Matrix.solutionByGaussMethod(matrix3x4, vector3f4);
         });
@@ -262,8 +262,8 @@ public class MatrixTest {
         Assertions.assertThat(thrown2.getMessage()).isNotBlank();
         Assertions.assertThat(thrown2.getMessage()).isEqualTo("There are no solutions");
 
-        Matrix4d matrix4x  = new Matrix4d(new float[]{1, -1, 2, -3, 1, 4, -1, -2, 1, -4, 3, -2, 1, -8, 5, -2});
-        Vector4d vector4f = new Vector4d(new float[]{1, -2, -2, -2});
+        Matrix4f matrix4x  = new Matrix4f(new float[]{1, -1, 2, -3, 1, 4, -1, -2, 1, -4, 3, -2, 1, -8, 5, -2});
+        Vector4f vector4f = new Vector4f(new float[]{1, -2, -2, -2});
         Assertions.assertThat(Matrix.solutionByGaussMethod(matrix4x, vector4f).getVector()).
                 isEqualTo(new float[]{-8, 4, 8, 1});
     }
